Move toggle controls to container level
This is mostly so that we can control the sidebar outside of itself,
we want to be able to use the sidebar like a container for its
contents like so:

```js
<Sidebar>
  <ul className="nav">...</ul>
</Sidebar>
```

So when an item is clicked in the nav, we want to toggle the menu
so that the content can be seen. This would be harder to do if
the toggle state was inside the Sidebar itself.